+++ include/aspect/simulator.h
@@ -93,8 +93,11 @@ namespace aspect
   template <int dim>
   class NewtonHandler;
 
-  template <int dim>
-  class FreeSurfaceHandler;
+  namespace MeshDeformation
+  {
+    template <int dim>
+    class FreeSurfaceHandler;
+  }
 
   namespace internal
   {
@@ -1726,11 +1729,11 @@ namespace aspect
        * if we do not need the machinery for doing free surface stuff, we do
        * not even allocate it.
        */
-      std::shared_ptr<FreeSurfaceHandler<dim> > free_surface;
+      std::shared_ptr<MeshDeformation::FreeSurfaceHandler<dim> > free_surface;
 
       friend class boost::serialization::access;
       friend class SimulatorAccess<dim>;
-      friend class FreeSurfaceHandler<dim>;  // FreeSurfaceHandler needs access to the internals of the Simulator
+      friend class MeshDeformation::FreeSurfaceHandler<dim>;  // FreeSurfaceHandler needs access to the internals of the Simulator
       friend struct Parameters<dim>;
   };
 }
