+++ include/aspect/simulator_access.h
@@ -128,8 +128,12 @@ namespace aspect
     template <int dim> class Manager;
   }
 
+  namespace MeshDeformation
+  {
+    template <int dim> class FreeSurfaceHandler;
+  }
+
   template <int dim> class MeltHandler;
-  template <int dim> class FreeSurfaceHandler;
 
   template <int dim> class NewtonHandler;
 
@@ -719,7 +723,7 @@ namespace aspect
        * Return a reference to the free surface handler. This function will
        * throw an exception if no free surface is activated.
        */
-      const FreeSurfaceHandler<dim> &
+      const MeshDeformation::FreeSurfaceHandler<dim> &
       get_free_surface_handler () const;
 
       /**
