+++ tests/inner_core_growth.cc
@@ -1,0 +1,1 @@
+#include "../cookbooks/inner_core_convection/inner_core_assembly.cc"
