+++ tests/inner_core_growth.prm
@@ -1,0 +1,149 @@
+# A simple setup for convection the inner core. See the
+# manual, Section Cookbooks/Geophysical setups/Inner core convection 
+# for more information. Changing the Rayleigh number Ra 
+# and the resistance to phase change P will change the mode
+# of inner core convection between translation, plume convection
+# or no convection.
+# 
+# Setup and equations are given in: Deguen, Alboussiere, 
+# and Cardin (2013), Thermal convection in Earth’s inner core 
+# with phase change at its boundary. GJI, 194, 1310-133.
+
+set Additional shared libraries            = ./libinner_core_convection.so 
+ 
+set Dimension                              = 2
+set Use years in output instead of seconds = false
+set End time                               = 50
+set Output directory                       = output-inner_core_traction
+set Pressure normalization                 = no
+
+# The equations are non-dimensionalized, and all of the 
+# material properties are constant and set to one, except
+# for the density, which scales with temperature.
+subsection Material model
+  set Model name = inner core material
+
+  subsection Simple model
+    set Reference density             = 0
+    set Reference specific heat       = 1
+    set Reference temperature         = 0
+    set Thermal conductivity          = 1
+    set Thermal expansion coefficient = 1
+    set Viscosity                     = 1
+  end
+
+# The 'inner core material' model also contains a function that
+# represents the resistance to melting/freezing at the inner core
+# boundary. 
+# This phase boundary model combines a prescribed a normal 
+# velocity and normal stress. The relation between the two scales 
+# with the phase change number P, which is given in the 'Function
+# expression'. 
+# For P-->inifinity, the boundary is a free slip boundary, and for 
+# P-->0, the boundary is an open boundary (with zero normal stress). 
+
+# For time dependence, the varibale names have to have dim+1 entries,
+# and the last variable will be the time
+  subsection Inner core
+    subsection Phase change resistance function
+      set Variable names      = x,y,z
+      set Function expression = 1e-2     # <-- P
+    end
+  end
+end
+
+# The inner core is a sphere, and the non-dimensional radius 
+# is 1.
+subsection Geometry model
+  set Model name = sphere
+
+  subsection Sphere
+    set Radius  = 1
+  end
+end
+
+# We use a new boundary condition at the surface, which takes 
+# into account the effect of melting/freezing of material at 
+# the boundary of inner and outer core, and which is implemented
+# in the shared library. 
+subsection Model settings
+  set Zero velocity boundary indicators       = 
+  set Prescribed traction boundary indicators = surface: zero traction
+  set Free surface boundary indicators = surface
+  set Prescribed velocity boundary indicators =
+  set Fixed temperature boundary indicators   = surface
+end
+
+subsection Free surface
+  set List of model names = function, free surface
+
+  subsection Function
+    set Function constants = growth_start=0, growth_rate=50
+    set Function expression = if(t<growth_start,0,x/50); \
+                              if(t<growth_start,0,y/50)
+  end
+end
+
+# We use a constant radiogenic heating rate.  
+subsection Heating model
+  set List of model names = constant core heating
+
+  subsection Constant core heating
+    set Radiogenic heating rate = 6
+  end
+end
+
+# The (potential) temperature is set to zero at the outer boundary.
+subsection Boundary temperature model
+  set List of model names = spherical constant
+  subsection Spherical constant
+    set Inner temperature = 0.8
+    set Outer temperature = 0
+  end
+end
+
+subsection Initial temperature model
+  set Model name = spherical hexagonal perturbation
+end
+
+# The gravity has its maximum value at the boundary of inner and
+# outer core, and decreases approximately linearly to zero towards
+# the center of the core. 
+# The Rayleigh number used in the model is given by the magnitude
+# of the gravity at the inner core/outer core boundary.
+subsection Gravity model
+  set Model name = radial linear
+
+  subsection Radial linear
+    set Magnitude at surface = 2     # <-- Ra
+  end
+end
+
+
+subsection Mesh refinement
+  set Initial global refinement          = 3
+  set Initial adaptive refinement        = 0
+  set Strategy                           = minimum refinement function
+  set Time steps between mesh refinement = 0
+
+  subsection Minimum refinement function
+    set Variable names = depth, phi, theta
+    set Function expression = if(depth>0.1,if(depth>0.3,5,6),7)
+  end
+end
+
+
+subsection Postprocess
+  set List of postprocessors = visualization, velocity statistics, temperature statistics, heat flux statistics
+
+  subsection Visualization
+    set Output format                 = vtu
+    set Time between graphical output = 0
+    set Number of grouped files       = 0
+    set List of output variables      = strain rate, gravity, density, specific heat, heating
+  end
+
+  subsection Depth average
+    set Time between graphical output = 1e6
+  end
+end
