+++ tests/inner_core_growth/statistics
@@ -1,0 +1,20 @@
+# 1: Time step number
+# 2: Time (seconds)
+# 3: Time step size (seconds)
+# 4: Number of mesh cells
+# 5: Number of Stokes degrees of freedom
+# 6: Number of temperature degrees of freedom
+# 7: Iterations for temperature solver
+# 8: Iterations for Stokes solver
+# 9: Velocity iterations in Stokes preconditioner
+# 10: Schur complement iterations in Stokes preconditioner
+# 11: Visualization file name
+# 12: RMS velocity (m/s)
+# 13: Max. velocity (m/s)
+# 14: Minimal temperature (K)
+# 15: Average temperature (K)
+# 16: Maximal temperature (K)
+# 17: Average nondimensional temperature (K)
+# 18: Outward heat flux through boundary with indicator 0 ("top") (W)
+0 0.000000000000e+00 0.000000000000e+00 320 2963 1313   0 90 92 359 output-inner_core_growth/solution/solution-00000 3.81223020e-04 6.57221284e-04 0.00000000e+00 2.68298907e-01 8.00000000e-01 3.35373634e-01 5.56152503e+00 
+1 5.000000000000e+01 5.000000000000e+01 320 2963 1313 128 89 91 363 output-inner_core_growth/solution/solution-00001 1.16796925e-02 1.17942385e-02 0.00000000e+00 1.16818055e+01 5.84714864e+00 1.46022569e+01 7.35988046e+01
